SystemMaster
The primary purpose of the program  to be the execution of various PowerShell commands related to system configuration and management on a Windows system. The GUI provides a dropdown menu with different commands, and upon selecting a command and clicking the "Run Command" button, the corresponding PowerShell command is executed.